+++ src/ispd/arquivo/xml/IconicoXML.java
@@ -1,8 +1,8 @@
 package ispd.arquivo.xml;
 
-import ispd.arquivo.xml.modelBuilders.CloudQueueNetworkBuilder;
-import ispd.arquivo.xml.modelBuilders.GridBuilder;
-import ispd.arquivo.xml.modelBuilders.QueueNetworkBuilder;
+import ispd.arquivo.xml.models.builders.CloudQueueNetworkBuilder;
+import ispd.arquivo.xml.models.builders.GridBuilder;
+import ispd.arquivo.xml.models.builders.QueueNetworkBuilder;
 import ispd.gui.PickModelTypeDialog;
 import ispd.gui.iconico.Edge;
 import ispd.gui.iconico.Vertex;
